Split out hover sample debounce logic so it can be more easily used in other places
