+++ osu.Game/Graphics/UserInterface/HoverSounds.cs
@@ -5,11 +5,8 @@ using System.ComponentModel;
 using osu.Framework.Allocation;
 using osu.Framework.Audio;
 using osu.Framework.Audio.Sample;
-using osu.Framework.Bindables;
 using osu.Framework.Extensions;
 using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Input.Events;
 using osu.Game.Configuration;
 using osu.Framework.Utils;
 
@@ -19,18 +16,11 @@ namespace osu.Game.Graphics.UserInterface
     /// Adds hover sounds to a drawable.
     /// Does not draw anything.
     /// </summary>
-    public class HoverSounds : CompositeDrawable
+    public class HoverSounds : HoverSampleDebounceComponent
     {
         private SampleChannel sampleHover;
 
-        /// <summary>
-        /// Length of debounce for hover sound playback, in milliseconds.
-        /// </summary>
-        public double HoverDebounceTime { get; } = 20;
-
         protected readonly HoverSampleSet SampleSet;
-
-        private Bindable<double?> lastPlaybackTime;
 
         public HoverSounds(HoverSampleSet sampleSet = HoverSampleSet.Normal)
         {
@@ -41,27 +31,13 @@ namespace osu.Game.Graphics.UserInterface
         [BackgroundDependencyLoader]
         private void load(AudioManager audio, SessionStatics statics)
         {
-            lastPlaybackTime = statics.GetBindable<double?>(Static.LastHoverSoundPlaybackTime);
-
             sampleHover = audio.Samples.Get($@"UI/generic-hover{SampleSet.GetDescription()}");
         }
 
-        protected override bool OnHover(HoverEvent e)
+        public override void PlayHoverSample()
         {
-            if (sampleHover == null)
-                return false;
-
-            bool enoughTimePassedSinceLastPlayback = !lastPlaybackTime.Value.HasValue || Time.Current - lastPlaybackTime.Value >= HoverDebounceTime;
-
-            if (enoughTimePassedSinceLastPlayback)
-            {
-                sampleHover.Frequency.Value = 0.96 + RNG.NextDouble(0.08);
-                sampleHover.Play();
-
-                lastPlaybackTime.Value = Time.Current;
-            }
-
-            return false;
+            sampleHover.Frequency.Value = 0.96 + RNG.NextDouble(0.08);
+            sampleHover.Play();
         }
     }
